Moved the DomainsMiniCart to its own file
